refactor: split up routes from `server.ts` into individual files
