+++ packages/opencode/src/server/routes/project.ts
@@ -1,12 +1,13 @@
 import { Hono } from "hono"
 import { describeRoute, validator } from "hono-openapi"
 import { resolver } from "hono-openapi"
-import { Instance } from "../project/instance"
-import { Project } from "../project/project"
+import { Instance } from "../../project/instance"
+import { Project } from "../../project/project"
 import z from "zod"
-import { errors } from "./error"
+import { errors } from "../error"
+import { lazy } from "../../util/lazy"
 
-export const ProjectRoute = new Hono()
+export const ProjectRoutes = lazy(() => new Hono()
   .get(
     "/",
     describeRoute({
@@ -77,3 +78,4 @@ export const ProjectRoute = new Hono()
       return c.json(project)
     },
   )
+)
