+++ packages/opencode/src/server/routes/question.ts
@@ -1,11 +1,12 @@
 import { Hono } from "hono"
 import { describeRoute, validator } from "hono-openapi"
 import { resolver } from "hono-openapi"
-import { Question } from "../question"
+import { Question } from "../../question"
 import z from "zod"
-import { errors } from "./error"
+import { errors } from "../error"
+import { lazy } from "../../util/lazy"
 
-export const QuestionRoute = new Hono()
+export const QuestionRoutes = lazy(() => new Hono()
   .get(
     "/",
     describeRoute({
@@ -93,3 +94,4 @@ export const QuestionRoute = new Hono()
       return c.json(true)
     },
   )
+)
